+++ crates/motsu/src/lib.rs
@@ -3,6 +3,7 @@
 extern crate alloc;
 
 mod context;
+mod precompiles;
 pub mod prelude;
 mod revert;
 mod router;