+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ### Added
 
+- Predeploy `EcRecover` precompile contract. #90
 - Support `fallback` and `receive` functions. #88
 - Derived `PartialEq`, `Eq` and `Debug` for `Account`. #81
 - Enabled additional conversions between `PrivateKeySigner`, `Address` and `Account`. #84
@@ -18,6 +19,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 ### Changed (Breaking)
 
 - Missing selector returns an error instead of panicking. #88
+- `evm::gas_left` and `evm::ink_left` return maximum values. #90
 
 ### Changed